fixed a couple of integer size warnings (MSVC)

+++ lib/tre-match-utils.h
@@ -64,7 +64,7 @@
 		  }							      \
 		else							      \
 		  {							      \
-		    pos_add_next = w;					      \
+		    pos_add_next = (unsigned int)w;					      \
 		    str_byte += w;					      \
 		  }							      \
 	      }								      \
